+++ lucid/nn/modules/rnn.py
@@ -4,7 +4,7 @@ Recurrent modules: LSTM, GRU, RNN.
 
 import math
 from collections import OrderedDict
-from typing import TYPE_CHECKING, cast, override
+from typing import cast, override
 
 from lucid._tensor.tensor import Tensor
 from lucid._types import DeviceLike, DTypeLike